v0.6.5
Fixed the DNC test at the end of setting up missing settings; turns out
Finder was not finding existing files/folders so soon after creating
them, so the fix is to pause the program for a couple of seconds.

Binary in the install folder is now a compressed file for easy
downloading.