v0.7.7-1
SSH-Check now runs iconfigSSH.py inside itself... so if the copy stored
in the ~/.ssh-check is out of date, it will not cacause any bugs in
SSH-Check. Ideally, if it is out of date the program should detect and
replace it, but that is for a furture feature. All you have to do is
delete python config file in .ssh-check and it will refresh with the
newest copy.